CDK/requirements.txt
aws-cdk.core>=1.132.0
aws-cdk.assertions>=1.132.0
aws-cdk.aws-ec2>=1.132.0
